Fix rare crash when deleting a node
Fix #1351

+++ gns3/items/node_item.py
@@ -149,6 +149,8 @@ class NodeItem():
         :param node_id: node identifier (integer)
         """
 
+        if self is None:
+            return
         self._initialized = True
         self.update()
         self._showLabel()
@@ -159,6 +161,8 @@ class NodeItem():
         when a the node has started.
         """
 
+        if self is None:
+            return
         for link in self._links:
             link.update()
 
@@ -168,6 +172,8 @@ class NodeItem():
         when a the node has stopped.
         """
 
+        if self is None:
+            return
         for link in self._links:
             link.update()
 
@@ -177,6 +183,8 @@ class NodeItem():
         when a the node has suspended.
         """
 
+        if self is None:
+            return
         for link in self._links:
             link.update()
 
@@ -205,6 +213,8 @@ class NodeItem():
         when a all the links must be deleted.
         """
 
+        if self is None:
+            return
         for link in self._links.copy():
             link.delete()
